RowBounds追加

+++ src/main/java/com/example/demo/domain/service/users/UsersService.java
@@ -18,7 +18,7 @@ public interface UsersService {
 	 */
 	public List<UsersBean> find(Long organizationsId,
             String name, Boolean deleted, String orderByClause,
-            Integer limit, Integer offset);
+            Long limit, Long offset);
 
 	/**
 	 * ユーザ取得
